requirements.txt
fastapi
requests
uvicorn
PyPDF2
sentence-transformers
torch
faiss-cpu
scikit-learn
langchain
langchain-community
langchain-openai
openai
python-multipart
elevenlabs
python-dotenv
easyocr
pymupdf
opencv-python-headless
pyttsx3
pdfplumber
pydantic